Client mode for Classic Paxos objects.
Added client operations to ClassicPaxos type, like Propose, which hide
the RPC details. With this change, it is not mandatory for an objects
to be configured in one of proposer, acceptor or learner roles. They
can be used as a client if they don't take any role.

Also, fixed some bugs and did some refactoring.

Reviewers        : bvk
Tests Run        :
Tickets Resolved :